osc: Modified Proportions and other stuff
Timecodes are now smaller which allowed to make the whole OSC slimmer.
The title that gets printed when the title in the OSC is clicked will now print the playlist position and length if there is a playlist with more than one entry.
OSC gets now re-init-ed when the file changes.